+++ packages/react/src/shared/createReactProxyDecoration.ts
@@ -1,0 +1,33 @@
+import { AnyRouter } from '@trpc/server';
+import { createProxy } from '@trpc/server/shared';
+import { CreateReactQueryHooks } from '../createReactQueryHooks';
+import { getQueryKey } from '../internals/getQueryKey';
+
+/**
+ * Create proxy for decorating procedures
+ * @internal
+ */
+export function createReactProxyDecoration<
+  TRouter extends AnyRouter,
+  TSSRContext = unknown,
+>(name: string, hooks: CreateReactQueryHooks<TRouter, TSSRContext>) {
+  return createProxy((opts) => {
+    const args = opts.args;
+
+    const pathCopy = [name, ...opts.path];
+
+    // The last arg is for instance `.useMutation` or `.useQuery()`
+    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+    const lastArg = pathCopy.pop()!;
+
+    // The `path` ends up being something like `post.byId`
+    const path = pathCopy.join('.');
+    if (lastArg === 'useMutation') {
+      return (hooks as any)[lastArg](path, ...args);
+    }
+    const [input, ...rest] = args;
+
+    const queryKey = getQueryKey(path, input);
+    return (hooks as any)[lastArg](queryKey, ...rest);
+  });
+}
